Improve gpadl message error handling
